+++ CHANGELOG.md
@@ -6,7 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Update dependency tools.jackson:jackson-bom to v3.1.2
+- Make Jackson an opt-in dependency
+- Support Jackson 2 and Jackson 3
 - Add `module-info.java`
 - Use JSpecify for nullability