+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.5.1] - 2026-06-23
 ### Fixed
 - Added `Json.instance()` to obtain an instance using SPI.
 
@@ -47,7 +49,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Java implementation ([#1](https://github.com/cucumber/messages-ndjson/pull/1))
 
-[Unreleased]: https://github.com/cucumber/messages-ndjson/compare/v0.5.0...HEAD
+[Unreleased]: https://github.com/cucumber/messages-ndjson/compare/v0.5.1...HEAD
+[0.5.1]: https://github.com/cucumber/messages-ndjson/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/cucumber/messages-ndjson/compare/v0.4.1...v0.5.0
 [0.4.1]: https://github.com/cucumber/messages-ndjson/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/cucumber/messages-ndjson/compare/v0.3.0...v0.4.0
@@ -56,4 +59,3 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.3.0]: https://github.com/cucumber/messages-ndjson/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/cucumber/messages-ndjson/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/cucumber/messages-ndjson/compare/210b25cf78776dc9e08a073d3e289e75402db885...v0.1.0
-dc9e08a073d3e289e75402db885...v0.1.0